Fixes for modern Mac guests
Network connectivity requires a specific NIC.

And since none of the built-in folder-sharing options work, we
need to assume sshfs is present, and use that. This is a big hack,
and requires the host to run sshd, and we need to share the user's
private ssh key with the guest, which isn't pretty, but it works.