Fix environment variable expansion
After the last changes, expanding the private key for private
boxes was broken, as it remained in the box. With this change,
it gets removed and moved into the URL before other environment
variables are expanded.

To test this, expand the set of mocking classes quite a bit so
that we can actually execute the bulk of the Vagrant file, and
capture some of the settings made for the virtual machines.