Proper PowerShell provisioning for mingw32
Rather than inserting a disk, install mingw via chocolatey. The
script is fairly generic, and accepts package name and version
overrides via role parameters.

To support this, add the relevant support to the provisioning
generator in the Vagrantfile.

The generic windows-tools role becomes a bit simplier so that it
is also useful for mingw. cmake, jom, and llvm are not strictly
needed by MSVC either anyway.