Fix discovery of winrm after upgrade to vagrant 2.2.5
Previously, we tried to load WinRM for each Windows box, and
selected the winrm communicator when this succeeded. With the
latest upgrade of vagrant, only the first loading attempt
succeeded, all others failed but without throwing an exception.
So we incorrectly fell back to the winssh communicator, but
without the corresponding message.

With this change, the has_winrm variable becomes global, and we
only try to load the winrm gem if it's false.