Modularize simple provisioning mechanisms into separate ruby files
Default vagrant provisioners are nice to be able to use directly
from the box definition without having to make a dedicated role.
By treating them structurally like other roles, we can keep the
main Vagrantfile somewhat clean from special case handling.

Only special case is scripts, which might use environment variables
that should not be expanded on the host, but on the guest. Those we
need to allow to not be expanded, and so we use $$ for those.