Capture output from psexec command
Redirect output into file, and print that file as it is generated using
tail. Abort the tail'ing when the "cmd exited" line is generated by
psexec, indicating that the command has finsihed.

This works well, probably best to unify the gui/non-gui case and always
use psexec, which gets rid of the error-code hack.