Make mapping paths the responsibility of the box
Doing in on the host means that multi-box runs will incorrectly
map $PWD to the wrong guest path, except for the first box.
Removing this logic from the host, and putting it into parse-opts
instead makes this just as simple for the job scripts, and removes
a lot of complexity from the run-on script.