Add support for Azure VMs
Requires the azure vagrant plugin, the azure dummy-box, the azure CLI,
and mutagen to be installed on the host machine.

Interpret the "box" as the Azure image.

Folder sharing would default to SMB, but hosts are usually not reachable
from Azure,
so we use mutagen to sync the minicoin folder to the VM.