Rewrite run-logic for Windows to use guest-side powershell
This simplifies the host-side script a lot, and makes it unnecessary to use
the network share for sending job output to the host. Instead, use powershell
to run the psexec process in the background, and "tail" files locally.

Since powershell doesn't have a clean concept of stderr, this is a bit more
messy than expected.

For this to work on Azure machines, username/password need to become
parameters to the script.