Explicitly specify the ssh forwarding port for each machine
This way, the port used by a suspended or shut-down machine
won't be reused by vagrant when a new machine is brought up.

This prevents that an SSH key is stored in known_hosts that no longer
matches the port.

If the list of boxes changes, then the port will change, so this is not a
100% solution. But it works for the 99% that matter.