Install VC++2019 through a job, and keep it alive
The Azure Load Balancer has a default idle timeout of 4 minutes. If there is
no traffic, then the connection dies, making vagrant wait for output.

Since the installer runs for longer than that, we need to keep the connection
alive through a heartbeat signal. Use a background job to run the
installer, and wait for that job while keeping both output and heartbeat
signal going over the network.