Remove overly complex continuous run support
The benefits are marginal, and don't justify the complexity and fragility
esp when interrupting a running job.

A better solution would perhaps implement the wait-loop on the guest
anyway, either simply running in a loop until the job is removed, or
waiting for a trigger (which then again requires costly communication).