Create machine definition from coin definitions
If COIN_ROOT points to the coin directory, read configuration files and
define machines based on them. For, read only the cmake_platforms.yaml
file from coin, and ignore "Includes".

coin template are mapped to vagrant boxes based on a map in the global
minicoin.yml.

Configure and build settings from the coin config are for now dumped
into the machine definition as "jobs". In the future, the minicoin "run"
command can perhaps read those and translate it into parameters.

Add resolution of internal hosts to the provisioning of the coin-node
role.