Add ccache support
The build job configures with -ccache by default if ccache is installed.
Implies -no-pch, otherwise ccache won't work.

Add ccache role, which installs ccache (supporting only apt and choco for
now), and configures the cache directory to whatever is passed in. This
should be turned on using a default minicoin role such as

settings:
  defaults:
    roles:
      - role: ccache
        cache_dir: $GUEST_HOMES/$USER/.ccache

which then results in the cache living on the host, ie. being persistent
and shared across machines.