Better validation of Azure requirements
Since we can't define Azure specific provisioners outside of the
provider-specific configuration block, they always come last. This is
why we depend on dependency_provisioners, which is experimental.

Setting that environment after vagrant has launched stopped working,
so instead, use a local_command provisioner for all boxes to validate
the requirements. If the validation fails, abort.

This way, the box boots, but will not get provisioned until things are
fixed.