Support for machine-defined job configurations
Replaces the experimental implementation made for coin support with a
cleaner solution for finding and selecting job configurations defined
by a machine.

This allows a machine to define which parameters a certain job should
receive by default. Ie. a machine that is provisioned with a certain SDK
can provide a jobconfig that uses that SDK.

Job configs are matched by name. If several jobconfigs match the
requested name (or if no name is specified) then list them and ask the
user to pick one.