Rewrite mutagen support as a synced_folder plugin
This removes a number of hacks from the provisioner implementation, as
we have more access to Vagrant facilities in the plugin.

The mutagen commands are now done immediately after boot, and before
regular folder syncing.

The action hook interface is not optimal, as the hooks are getting
called a lot of times. Need to find a better way for that.