Better job output handling
Support pattern matchers that specify a color or don't want a newline.
Since a line that wants a newline following a line that doens't means
that there should be a newline between them, store the running previous
option separately.

This no-newline support only applies in sequential runs.

Add a manual test case.