Ubuntu 20.04 on AWS
The box is based on the default AMI from the market place, the vagrant
bootstrapping is done via user data.

Replacing the insecure SSH key that is used for the first connection
kills the connection, so not replacing it for now. Ideally we would
bootstrap using a dedicated key that is generated on the host, but then
we need to put it in the user data as well which needs some more work.

"Implement" the cloud folder synching with mutagen, it all just works.

Include an AWS role that allows minicoin machines to override settings.