AWS: Implement support for Windows machines
Add a Windows2019 based box that is bootstrapped to support vagrant and
access via both WinRM and SSH. Add AWS specific configurations to the
general AWS provider configuration.

Authentication uses the insecure vagrant key for now, which needs to be
fixed later (also for the Ubuntu machine) to use the user's public key.