Clean up AWS support implementation
Move common code to general provider configuration so that boxes only
have to provide the AMI and the user data.

Passing the admin password for the box through to the box-specific
override is messy, but since all boxes of the same provider will have
the same password, we can use an environment variable.
To pass the admin password through the cloud_prepare plugin we can use
folder options.