Sync the minicoin folder to /opt/minicoin instead of /minicoin
The macOS root file system is read only, which we could work around on a
local VM by adding a link to /etc/synthetic.conf, but it's both hacky
and not feasibly for macOS machines living in the cloud. And adding stuff
to / is also awkward on Linux.

So move things to /opt instead, which on Windows because c:\opt\minicoin
which isn't any more weird that c:\minicoin. Adjust all the job scripts
and the internal explicit handling of the default shared folder.