AWS: Move user_data scripts into separate file
That file can then be loaded from the generic AWS code, so we don't need
to package a new box whenever a modification is needed. Instead, move the
instance and EBS configuration back into each box file so that we don't
need platform-specific logic in the generic code.

With this change, we no longer need to pass data around through global
variables.