Clean up AWS and Azure boxes and plumbing
Those boxes can be published on the vagrant cloud, so we don't need to
maintain our own metadata files. The 'minicoin' organization on the
vagrant cloud is the namespace for the boxes when published, so adjust
workaround accordingly.

Add an info.json file with some more meta data.

Consolidate the box packaging script and add a script for publishing
everything to vagrant.