Mutagen: better error handling if there is no default SSH key
Mutagen doesn't allow us to specify the private key to use, so we
have to make sure that the default key works. If there is no default
key, mutagen would fail, so raise an exception.

The shared_folder logic already handles that in the usable?
implementation, this covers the create command as well.