Cleanup: get rid of reverse mutagen mode
It's never been implemented for all platform, is completely untested,
useless for cloud machines, and overly complex. With the current version
of mutagen, we also don't need to create the directory structure on the
guest anymore, that shaves a few more seconds from creating a machine.

We can assume that mutagen is on the host, or that the configuration
doesn't use mutagen.