Start turning the mutagen role into a shared folder type
It feels odd to use a "role" for setting up folder sharing when there
already is a shared_folders setting. So instead, allow setting up of
mutagen folder sharing in the shared_folders configuration, by setting
a "type" on the alpha:beta mapping.

The only disadvantage is that we can't just list a bunch of paths, as
each path needs to have the type set.