Windows: jobconfigs for multiple compilers on the same VM
Create jobconfigs for the build job for VC++ and MinGW that apply
when building Qt (or any project with a configure script). If both
are provisioned on a machine, then running the build job will ask
for which compiler to use (unless explicitly specified).

To build each jobconfig into its own build tree, interpret a build
path parameter that starts with a dash as a suffix, so the build tree
becomes e.g. dev-build-msvc.