Add a command "machine" for managing machine files
Until now, all machines are declared in one of the minicoin.yml files.
This makes the list of machines very, and thus minicoin very slow, if a
lot of rarely used machines are added. For example, few people use the
android machines regularly.

Make it possible to declare a machine in a separate yaml file, located
in the "machines" directory. The library are those machine files shipped
as part of minicoin, and machine files can with the "machine" command be
copied to the user's minicoin directory, from where they are loaded.

Customization of those machines then happens as before in the user's
minicoin.yml file, or a project-local file. Or, since the machine files
get copied, users can of course customise them there.