Refactor role implementation and apply roles in order
We need to first load all roles and follow the dependencies and
aliases to create a correctly ordered list of roles. Then we can
apply each role in order.

Otherwise a "qtspeech-builder" role might be applied before the
"qt-builder" role is applied, if both roles are explicitly listed
out-of-order (perhaps one from an installed machine, the other from
a user setting).